+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,40 @@
 # alpha-test-app
 
+## 0.1.0-alpha.140
+
+### Patch Changes
+
+- Updated dependencies [02f5691]
+  - sit-onyx@1.0.0-alpha.143
+
+## 0.1.0-alpha.139
+
+### Patch Changes
+
+- Updated dependencies [7c98a6d]
+  - sit-onyx@1.0.0-alpha.142
+
+## 0.1.0-alpha.138
+
+### Patch Changes
+
+- Updated dependencies [08b434b]
+  - sit-onyx@1.0.0-alpha.141
+
+## 0.1.0-alpha.137
+
+### Patch Changes
+
+- Updated dependencies [d3e9321]
+  - sit-onyx@1.0.0-alpha.140
+
+## 0.1.0-alpha.136
+
+### Patch Changes
+
+- Updated dependencies [107ec36]
+  - sit-onyx@1.0.0-alpha.139
+
 ## 0.1.0-alpha.135
 
 ### Patch Changes