+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,40 @@
 # playground
 
+## 0.0.1-alpha.82
+
+### Patch Changes
+
+- Updated dependencies [02f5691]
+  - sit-onyx@1.0.0-alpha.143
+
+## 0.0.1-alpha.81
+
+### Patch Changes
+
+- Updated dependencies [7c98a6d]
+  - sit-onyx@1.0.0-alpha.142
+
+## 0.0.1-alpha.80
+
+### Patch Changes
+
+- Updated dependencies [08b434b]
+  - sit-onyx@1.0.0-alpha.141
+
+## 0.0.1-alpha.79
+
+### Patch Changes
+
+- Updated dependencies [d3e9321]
+  - sit-onyx@1.0.0-alpha.140
+
+## 0.0.1-alpha.78
+
+### Patch Changes
+
+- Updated dependencies [107ec36]
+  - sit-onyx@1.0.0-alpha.139
+
 ## 0.0.1-alpha.77
 
 ### Patch Changes