+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @sit-onyx/nuxt
 
+## 1.0.0-alpha.23
+
+### Patch Changes
+
+- Updated dependencies [02f5691]
+  - sit-onyx@1.0.0-alpha.143
+
+## 1.0.0-alpha.22
+
+### Patch Changes
+
+- Updated dependencies [7c98a6d]
+  - sit-onyx@1.0.0-alpha.142
+
+## 1.0.0-alpha.21
+
+### Patch Changes
+
+- Updated dependencies [08b434b]
+  - sit-onyx@1.0.0-alpha.141
+
+## 1.0.0-alpha.20
+
+### Patch Changes
+
+- Updated dependencies [d3e9321]
+  - sit-onyx@1.0.0-alpha.140
+
+## 1.0.0-alpha.19
+
+### Patch Changes
+
+- Updated dependencies [107ec36]
+  - sit-onyx@1.0.0-alpha.139
+
 ## 1.0.0-alpha.18
 
 ### Patch Changes