+++ packages/storybook-utils/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @sit-onyx/storybook-utils
 
+## 1.0.0-alpha.148
+
+### Patch Changes
+
+- Updated dependencies [02f5691]
+  - sit-onyx@1.0.0-alpha.143
+
+## 1.0.0-alpha.147
+
+### Patch Changes
+
+- Updated dependencies [7c98a6d]
+  - sit-onyx@1.0.0-alpha.142
+
+## 1.0.0-alpha.146
+
+### Patch Changes
+
+- Updated dependencies [08b434b]
+  - sit-onyx@1.0.0-alpha.141
+
+## 1.0.0-alpha.145
+
+### Patch Changes
+
+- Updated dependencies [d3e9321]
+  - sit-onyx@1.0.0-alpha.140
+
+## 1.0.0-alpha.144
+
+### Patch Changes
+
+- Updated dependencies [107ec36]
+  - sit-onyx@1.0.0-alpha.139
+
 ## 1.0.0-alpha.143
 
 ### Patch Changes